+++ packages/f-error-message/src/components/ErrorMessage.vue
@@ -4,20 +4,22 @@
         :class="$style['c-errorMessage']"
         data-test-id="error-message"
         :data-test-title="dataTestTitle">
-        <warning-icon :class="$style['c-errorMessage-icon']" />
-        <span data-test-id="content">
+        <danger-icon :class="$style['c-errorMessage-icon']" />
+        <span
+            data-test-id="content"
+            :class="$style['c-errorMessage-content']">
             <slot />
         </span>
     </p>
 </template>
 
 <script>
-import { WarningIcon } from '@justeat/f-vue-icons';
+import { DangerIcon } from '@justeat/f-vue-icons';
 
 export default {
     name: 'ErrorMessage',
     components: {
-        WarningIcon
+        DangerIcon
     },
     props: {
         dataTestTitle: {
@@ -34,19 +36,22 @@ export default {
 </script>
 
 <style lang="scss" module>
-
 .c-errorMessage {
-    display: flex;
-    align-items: center;
+    position: relative;
     color: $color-text--danger;
     @include font-size(base);
     margin-top: spacing();
 }
 
-.c-errorMessage-icon {
-    width: 16px;
-    height: 16px;
-    margin-right: spacing(x0.5);
+.c-errorMessage-content {
+    display: block;
+    overflow: hidden;
+    margin-left: spacing(x3);
 }
 
+.c-errorMessage-icon {
+    position: absolute;
+    width: 16px;
+    margin-top: 2px;
+}
 </style>