+++ packages/f-error-message/CHANGELOG.md
@@ -3,6 +3,16 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.3.0
+------------------------------
+*November 12, 2020*
+
+### Changed
+- Icon from `Warning Icon` to `Danger Icon`.
+
+### Added
+- Styling to display icon inline with text for multiline messages.
+
 
 v0.2.0
 ------------------------------
@@ -11,6 +21,7 @@ v0.2.0
 ### Added
 - Built out functionality
 - Added unit tests
+
 
 v0.1.0
 ------------------------------